requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
requests==2.31.0
flask==2.0.1
APScheduler==3.6.3 